requirements.txt
python-dotenv
web3
python-telegram-bot